+++ src/main/java/org/jboss/forge/arquillian/api/algeron/AlgeronRetrieverFacet.java
@@ -6,9 +6,9 @@ import org.jboss.forge.addon.facets.constraints.FacetConstraint;
 import org.jboss.forge.addon.projects.Project;
 import org.jboss.forge.addon.projects.ProjectFacet;
 import org.jboss.forge.addon.projects.facets.DependencyFacet;
-import org.jboss.forge.arquillian.api.ArquillianConfig;
+import org.jboss.forge.arquillian.model.core.ArquillianConfig;
 import org.jboss.forge.arquillian.api.core.ArquillianFacet;
-import org.jboss.forge.arquillian.api.YamlGenerator;
+import org.jboss.forge.arquillian.util.YamlGenerator;
 
 import java.util.Map;
