+++ src/main/java/org/jboss/forge/arquillian/api/core/ArquillianExtensionFacet.java
@@ -6,7 +6,7 @@ import org.jboss.forge.addon.facets.constraints.FacetConstraints;
 import org.jboss.forge.addon.projects.ProjectFacet;
 import org.jboss.forge.addon.projects.facets.DependencyFacet;
 import org.jboss.forge.addon.projects.facets.MetadataFacet;
-import org.jboss.forge.arquillian.api.Extension;
+import org.jboss.forge.arquillian.extension.Extension;
 import org.jboss.forge.arquillian.extension.ExtensionResolver;
 
 import javax.inject.Inject;
