+++ src/main/java/org/jboss/forge/arquillian/api/core/testframework/JUnitFacet.java
@@ -4,41 +4,39 @@
  * Licensed under the Eclipse Public License version 1.0, available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.jboss.forge.arquillian.testframework;
+package org.jboss.forge.arquillian.api.core.testframework;
 
 import org.jboss.forge.addon.dependencies.builder.DependencyBuilder;
-import org.jboss.forge.arquillian.api.core.TestFrameworkFacet;
 
 /**
  * @author <a href="mailto:bartosz.majsak@gmail.com">Bartosz Majsak</a>
  */
-public class TestNGFacet extends TestFrameworkFacet {
+public class JUnitFacet extends TestFrameworkFacet {
     @Override
     public String getFrameworkName() {
-        return "testng";
+        return "junit";
     }
 
     @Override
     public String getTemplateLocation() {
-        return "/TestNGTest.ftl";
+        return "/JUnitTest.ftl";
     }
 
     @Override
     public String getTemplateStandaloneLocation() {
-        return "/TestNGStandaloneTest.ftl";
+        return "/JUnitStandaloneTest.ftl";
     }
 
     @Override
     public String getVersionPropertyName() {
-        return "version.testng";
+        return "version.junit";
     }
-
 
     @Override
     public DependencyBuilder createFrameworkDependency() {
         return DependencyBuilder.create()
-            .setGroupId("org.testng")
-            .setArtifactId("testng")
+            .setGroupId("junit")
+            .setArtifactId("junit")
             .setScopeType("test");
     }
 
@@ -46,7 +44,7 @@ public class TestNGFacet extends TestFrameworkFacet {
     public DependencyBuilder createArquillianDependency() {
         return DependencyBuilder.create()
             .setGroupId("org.arquillian.universe")
-            .setArtifactId("arquillian-testng")
+            .setArtifactId("arquillian-junit")
             .setScopeType("test")
             .setPackaging("pom");
     }
@@ -55,7 +53,7 @@ public class TestNGFacet extends TestFrameworkFacet {
     public DependencyBuilder createArquillianStandaloneDependency() {
         return DependencyBuilder.create()
             .setGroupId("org.arquillian.universe")
-            .setArtifactId("arquillian-testng-standalone")
+            .setArtifactId("arquillian-junit-standalone")
             .setScopeType("test")
             .setPackaging("pom");
     }
