+++ src/main/java/org/jboss/forge/arquillian/command/algeron/provider/AlgeronSetupProviderCommand.java
@@ -12,7 +12,7 @@ import org.jboss.forge.addon.ui.result.Result;
 import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
-import org.jboss.forge.arquillian.api.core.TestFrameworkFacet;
+import org.jboss.forge.arquillian.api.core.testframework.TestFrameworkFacet;
 import org.jboss.forge.arquillian.command.algeron.AbstractAlgeronCommand;
 import org.jboss.forge.arquillian.model.algeron.ContractProviderLibrary;
 import org.jboss.forge.arquillian.api.algeron.AlgeronProviderFacet;
