+++ src/main/java/org/jboss/forge/arquillian/command/core/ContainerConfigurationCommand.java
@@ -16,7 +16,7 @@ import org.jboss.forge.addon.ui.result.Result;
 import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
-import org.jboss.forge.arquillian.api.ArquillianConfig;
+import org.jboss.forge.arquillian.model.core.ArquillianConfig;
 import org.jboss.forge.arquillian.api.core.ArquillianFacet;
 import org.jboss.forge.arquillian.container.ProfileManager;
 import org.jboss.forge.arquillian.container.model.Configuration;
