+++ src/main/java/org/jboss/forge/arquillian/command/cube/CubeSetupCommand.java
@@ -21,8 +21,8 @@ import org.jboss.forge.addon.ui.result.Results;
 import org.jboss.forge.addon.ui.util.Categories;
 import org.jboss.forge.addon.ui.util.Metadata;
 import org.jboss.forge.arquillian.api.core.ArquillianFacet;
-import org.jboss.forge.arquillian.api.core.TestFrameworkFacet;
-import org.jboss.forge.arquillian.api.YamlGenerator;
+import org.jboss.forge.arquillian.api.core.testframework.TestFrameworkFacet;
+import org.jboss.forge.arquillian.util.YamlGenerator;
 import org.jboss.forge.arquillian.api.cube.CubeSetupFacet;
 import org.jboss.forge.arquillian.model.cube.CubeConfiguration;
 import org.jboss.forge.furnace.util.Strings;
