+++ src/main/java/org/jboss/forge/arquillian/container/ContainerInstallEvent.java
@@ -4,7 +4,7 @@
  * Licensed under the Eclipse Public License version 1.0, available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.jboss.forge.arquillian.api;
+package org.jboss.forge.arquillian.container;
 
 import org.jboss.forge.arquillian.container.model.Container;
