+++ src/main/java/org/jboss/forge/arquillian/container/provider/ContainerDirectoryParser.java
@@ -4,7 +4,7 @@
  * Licensed under the Eclipse Public License version 1.0, available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.jboss.forge.arquillian.container.index;
+package org.jboss.forge.arquillian.container.provider;
 
 import org.codehaus.jackson.map.ObjectMapper;
 import org.codehaus.jackson.map.PropertyNamingStrategy;
