+++ src/main/java/org/jboss/forge/arquillian/container/provider/ContainerIndexLocationProvider.java
@@ -4,7 +4,7 @@
  * Licensed under the Eclipse Public License version 1.0, available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.jboss.forge.arquillian.container.index;
+package org.jboss.forge.arquillian.container.provider;
 
 import java.net.URL;
