+++ src/main/java/org/jboss/forge/arquillian/container/provider/FileContainerIndexLocationProvider.java
@@ -4,22 +4,21 @@
  * Licensed under the Eclipse Public License version 1.0, available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.jboss.forge.arquillian.container.index;
+package org.jboss.forge.arquillian.container.provider;
 
-import javax.enterprise.inject.Alternative;
 import java.net.MalformedURLException;
+import java.net.URISyntaxException;
 import java.net.URL;
 
 /**
  * @Author Paul Bakker - paul.bakker.nl@gmail.com
  */
-@Alternative
-public class DocumentationContainerIndexLocationProvider implements ContainerIndexLocationProvider {
+public class FileContainerIndexLocationProvider implements ContainerIndexLocationProvider {
     @Override
     public URL getUrl() {
         try {
-            return new URL("https://raw.github.com/gist/1324966/afe53313a2ed345585188a5c1f3d43fdb0c667d6/containers.json");
-        } catch (MalformedURLException e) {
+            return this.getClass().getClassLoader().getResource("containers.json").toURI().toURL();
+        } catch (MalformedURLException | URISyntaxException e) {
             throw new RuntimeException(e);
         }
     }
