+++ src/main/java/org/jboss/forge/arquillian/extension/Extension.java
@@ -1,4 +1,4 @@
-package org.jboss.forge.arquillian.api;
+package org.jboss.forge.arquillian.extension;
 
 import org.jboss.forge.addon.dependencies.Dependency;
