+++ src/main/java/org/jboss/forge/arquillian/extension/ExtensionResolver.java
@@ -10,7 +10,6 @@ import org.jboss.forge.addon.maven.resources.MavenModelResource;
 import org.jboss.forge.addon.projects.Project;
 import org.jboss.forge.addon.projects.facets.DependencyFacet;
 import org.jboss.forge.arquillian.api.core.ArquillianFacet;
-import org.jboss.forge.arquillian.api.Extension;
 
 import javax.inject.Inject;
 import java.io.File;
