+++ src/test/java/test/integration/algeron/AddArquillianAlgeronRetrieverTest.java
@@ -2,10 +2,9 @@ package test.integration.algeron;
 
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.forge.addon.resource.FileResource;
-import org.jboss.forge.arquillian.api.ArquillianConfig;
+import org.jboss.forge.arquillian.model.core.ArquillianConfig;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import test.integration.extension.AddDependencies;
 import test.integration.extension.AddPackage;
 import test.integration.support.ShellTestTemplate;
 import test.integration.support.assertions.ForgeAssertions;
